requirements.txt
numpy
opencv-python==4.5.3.56
flask
Flask-HTTPAuth
python-dotenv
Flask-Session
# gunicorn
# depthai